+++ tests/com/yanncebron/m68kplugin/lang/resolve/M68kLabelResolveTest.java
@@ -16,8 +16,13 @@
 
 package com.yanncebron.m68kplugin.lang.resolve;
 
+import com.intellij.codeInsight.daemon.impl.HighlightInfoType;
+import com.intellij.codeInsight.daemon.impl.IdentifierHighlighterPassFactory;
+import com.intellij.codeInsight.daemon.impl.SeveritiesProvider;
+import com.intellij.codeInsight.intention.IntentionAction;
 import com.intellij.codeInsight.lookup.LookupElement;
 import com.intellij.codeInsight.lookup.LookupElementPresentation;
+import com.intellij.codeInspection.ex.InspectionProfileImpl;
 import com.intellij.icons.AllIcons;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiPolyVariantReference;
@@ -25,14 +30,19 @@ import com.intellij.psi.PsiReference;
 import com.intellij.psi.ResolveResult;
 import com.intellij.testFramework.TestDataPath;
 import com.intellij.testFramework.fixtures.BasePlatformTestCase;
+import com.intellij.usages.Usage;
 import com.intellij.util.ArrayUtil;
 import com.intellij.util.containers.ContainerUtil;
 import com.intellij.util.xml.ElementPresentationManager;
 import com.yanncebron.m68kplugin.inspections.M68kUnresolvedLabelReferenceInspection;
-import icons.M68kIcons;
 import com.yanncebron.m68kplugin.lang.psi.M68kBuiltinSymbol;
 import com.yanncebron.m68kplugin.lang.psi.M68kBuiltinSymbolPsiElement;
 import com.yanncebron.m68kplugin.lang.psi.M68kLocalLabel;
+import icons.M68kIcons;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Collection;
+import java.util.List;
 
 import static com.yanncebron.m68kplugin.lang.M68kLookupElementTestUtil.*;
 
@@ -54,6 +64,12 @@ public class M68kLabelResolveTest extends BasePlatformTestCase {
       ContainerUtil.map2Array(M68kBuiltinSymbol.values(), String.class, M68kBuiltinSymbol::getName));
   }
 
+  private void setupImplicitMacroConfiguration() {
+    M68kUnresolvedLabelReferenceInspection inspection = new M68kUnresolvedLabelReferenceInspection();
+    inspection.labelDefiningMacros.add("implicitLabelMacro");
+    myFixture.enableInspections(inspection);
+  }
+
   public void testResolveBuiltinSymbol() {
     myFixture.configureByText("a.s",
       "MY_CPU equ __C<caret>PU");
@@ -64,17 +80,19 @@ public class M68kLabelResolveTest extends BasePlatformTestCase {
   }
 
   public void testHighlightResolvingInSingleFile() {
-    myFixture.enableInspections(new M68kUnresolvedLabelReferenceInspection());
+    setupImplicitMacroConfiguration();
     myFixture.testHighlighting("highlightResolvingInSingleFile.s");
   }
 
   public void testCompletionVariantsInSingleFile() {
+    setupImplicitMacroConfiguration();
     myFixture.testCompletionVariants("completionVariantsInSingleFile.s",
       getCompletionVariants(
         "topLevelLabel", "anotherTopLevel.Label",
         ".localLabel", ".localLabel2", ".localLabelWithColon",
         "localLabelDollar$", "localLabelDollarWithColon$",
-        "setLabel", "equLabel", "equalsLabel")
+        "setLabel", "equLabel", "equalsLabel",
+        "implicitLabel")
     );
 
     final LookupElement topLevelLabel1 = findLookupElement(myFixture, "topLevelLabel", 0);
@@ -136,6 +154,17 @@ public class M68kLabelResolveTest extends BasePlatformTestCase {
     assertTrue(builtinSymbolPresentation.isItemTextItalic());
     assertEquals(" Current PC", builtinSymbolPresentation.getTailText());
     assertEquals("any", builtinSymbolPresentation.getTypeText());
+
+    final LookupElement implicitLabel1 = findLookupElement(myFixture, "implicitLabel", 46);
+    final LookupElementPresentation implicitLabel1Presentation = LookupElementPresentation.renderElement(implicitLabel1);
+    assertLookupIcon(implicitLabel1Presentation, M68kIcons.LABEL_MACRO);
+    assertTrue(implicitLabel1Presentation.isItemTextBold());
+    assertEquals(" (implicitLabelMacro implicitLabel)", implicitLabel1Presentation.getTailText());
+    final LookupElement implicitLabel2 = findLookupElement(myFixture, "implicitLabel", 80);
+    final LookupElementPresentation implicitLabelPresentation2 = LookupElementPresentation.renderElement(implicitLabel2);
+    assertLookupIcon(implicitLabelPresentation2, M68kIcons.LABEL_MACRO);
+    assertTrue(implicitLabelPresentation2.isItemTextBold());
+    assertEquals(" (implicitLabelMacro implicitLabel)", implicitLabelPresentation2.getTailText());
   }
 
   public void testCompletionVariantsInParenthesizedExpression() {
@@ -153,7 +182,7 @@ public class M68kLabelResolveTest extends BasePlatformTestCase {
 
   public void testHighlightResolveInMultipleFiles() {
     final String[] testDataPaths = {"highlightResolvingInMultipleFiles.s", "highlightResolvingInMultipleFiles_other.s"};
-    myFixture.enableInspections(new M68kUnresolvedLabelReferenceInspection());
+    setupImplicitMacroConfiguration();
     myFixture.testHighlighting(testDataPaths);
 
     // resolve in current file (2nd declaration in _other.s)
@@ -162,14 +191,80 @@ public class M68kLabelResolveTest extends BasePlatformTestCase {
   }
 
   public void testCompletionVariantsInMultipleFiles() {
+    setupImplicitMacroConfiguration();
     myFixture.copyFileToProject("highlightResolvingInMultipleFiles_other.s");
     myFixture.testCompletionVariants("completionVariantsInMultipleFiles.s",
-      getCompletionVariants("topLevelLabel", "anotherTopLevelLabel", "otherLabel", "otherLabel2", "myLabel"));
+      getCompletionVariants("topLevelLabel", "anotherTopLevelLabel", "otherLabel", "otherLabel2", "myLabel", "implicitLabel"));
 
     final LookupElement otherLabel = findLookupElement(myFixture, "otherLabel");
     final LookupElementPresentation otherLabelPresentation = LookupElementPresentation.renderElement(otherLabel);
     assertFalse(otherLabelPresentation.isItemTextBold());
     assertEquals("highlightResolvingInMultipleFiles_other.s", otherLabelPresentation.getTypeText());
+
+    final LookupElement implicitLabel = findLookupElement(myFixture, "implicitLabel");
+    final LookupElementPresentation implicitLabelPresentation = LookupElementPresentation.renderElement(implicitLabel);
+    assertLookupIcon(implicitLabelPresentation, M68kIcons.LABEL_MACRO);
+    assertFalse(implicitLabelPresentation.isItemTextBold());
+    assertEquals(" (implicitLabelMacro implicitLabel)", implicitLabelPresentation.getTailText());
+    assertEquals("highlightResolvingInMultipleFiles_other.s", implicitLabelPresentation.getTypeText());
+  }
+
+  public void testImplicitMacroLabelMacroCallQuickFix() {
+    runImplicitMacroQuickFixTest(1);
+  }
+
+  // todo M68kUnresolvedLabelReferenceInspection.isSuitableAsImplicitMacro
+  public void _testImplicitMacroLabelMacroCallUnsuitable() {
+    runImplicitMacroQuickFixTest(0);
+  }
+
+  public void testImplicitMacroLabelLabelRefQuickFix() {
+    runImplicitMacroQuickFixTest(1);
+  }
+
+  public void testImplicitMacroLabelLabelRefMultipleQuickFixes() {
+    runImplicitMacroQuickFixTest(2);
+  }
+
+  public void testImplicitMacroLabelHighlightUsages() {
+    setupImplicitMacroConfiguration();
+    SeveritiesProvider.EP_NAME.getPoint().registerExtension(new SeveritiesProvider() {
+      @Override
+      public @NotNull List<HighlightInfoType> getSeveritiesHighlightInfoTypes() {
+        return List.of(HighlightInfoType.ELEMENT_UNDER_CARET_READ, HighlightInfoType.ELEMENT_UNDER_CARET_WRITE);
+      }
+    }, getTestRootDisposable());
+    IdentifierHighlighterPassFactory.doWithHighlightingEnabled(getProject(), getTestRootDisposable(), () -> {
+      myFixture.setReadEditorMarkupModel(true);
+      myFixture.configureByFile(getTestName(true) + ".s");
+      myFixture.checkHighlighting();
+    });
+  }
+
+  public void testImplicitMacroLabelFindUsages() {
+    setupImplicitMacroConfiguration();
+    Collection<Usage> usages = myFixture.testFindUsagesUsingAction(getTestName(true) + ".s");
+    assertSize(2, usages);
+    assertSameElements(ContainerUtil.map(usages, Object::toString), "8|MY_EQU |equ| |implicitLabel", "6|lea| |implicitLabel|(|a0|),|a1");
+  }
+
+  private void runImplicitMacroQuickFixTest(int quickFixCount) {
+    try {
+      InspectionProfileImpl.INIT_INSPECTIONS = true;
+      myFixture.enableInspections(new M68kUnresolvedLabelReferenceInspection());
+
+      myFixture.testHighlighting(getTestName(true) + ".s");
+
+      assertSize(quickFixCount, myFixture.filterAvailableIntentions("Add macro"));
+
+      if (quickFixCount != 0) {
+        IntentionAction intention = myFixture.findSingleIntention("Add macro 'implicitLabelMacro' as defining label");
+        myFixture.launchAction(intention);
+        myFixture.checkResultByFile(getTestName(true) + "_after.s");
+      }
+    } finally {
+      InspectionProfileImpl.INIT_INSPECTIONS = false;
+    }
   }
 
   private ResolveResult getSingleResolveResultAtCaret(String... testDataPaths) {